refactor: renamed old tests
